[GameServer] WSL 호환성 확보 및 이동/타임아웃 시스템 고도화
- WSL/GCC 환경에서의 컴파일 오류 방지를 위해 모든 소스 파일의 UTF-8 BOM 제거함.
- 포탈 이동 시 목적지 포탈 ID가 아닌 소스 포탈 ID를 참조하던 데이터 정합성 오류 수정함.
- 이동 연산 효율을 높이기 위해 직선 경로에 대해 'dtNavMeshQuery::raycast' 최적화(Fast-path) 적용함.
- 타임아웃 시스템을 토큰 기반 자가 갱신 구조로 개편하고, 'OnRecvPacket'에서 모든 패킷 수신 시 자동 연장되도록 처리함.
- 객체 설계 및 패킷 흐름을 종합한 시스템 가이드 통합 문서 생성함.

+++ Server/GameServer/GameSession.h
@@ -22,6 +22,6 @@ public:
 private:
     shared_ptr<PlayerCharacter> _playerRef;
     shared_ptr<JobQueue> _jobQueue;
-    atomic<uint32> _timeOutToken;
+    atomic<uint64> _timeOutToken;
 };
